fix arc sequence handling in enumerator
Array#slice! with one integer argument will remove the member at that position of the array.
what is actually wanted, according to the reference implementation [0], is to cut off everything starting at that position which can be achieved by supplying an endless range.

this fixes an encoding issue with german.dict (and possibly other dictionaries) as mormor would produce

    Abbehalten_Kabbehalten_VER:INF:NONÄ:NON:NEB
    Abbehalten_Kabbehalten_VER:PA2:NON<84>:NON:NEB
    Abbehaltende_A_SUB:AKK:SIN:FEM:ADJ:NON:NEB

instead of

    Abbehalten_Kabbehalten_VER:INF:NON
    Abbehalten_Kabbehalten_VER:PA2:NON
    Abbehaltende_A_SUB:AKK:SIN:FEM:ADJ

resolves #3.
closes #2.

[0]: https://github.com/morfologik/morfologik-stemming/blob/d7f1c571829562eaa4a16382860c1db42a63fe85/morfologik-fsa/src/main/java/morfologik/fsa/ByteSequenceIterator.java#L141